Scale down whenever possible if the worker is idempotent
Since we cannot assume workers were idempotent or not. We used to scale
down to minimum only when there was no processing happening. But since
our users told they want workers to scale down to a less number whenever
possible we have put this behind the "idempotent" flag to make it
possible for such workers to scale down from 30 to 25 to 22 to 19 to 10
to 0 instead of only 30 to 0.